Move internals into src dir
