+++ lib/src/adjacency_graphs.dart
