+++ lib/src/result.dart
